Changed type of start from int to long in RangeDataPool

+++ OpenBullet2/Models/Data/RangeDataPoolOptions.cs
@@ -2,7 +2,7 @@
 {
     public class RangeDataPoolOptions : DataPoolOptions
     {
-        public int Start { get; set; } = 0;
+        public long Start { get; set; } = 0;
         public int Amount { get; set; } = 100;
         public int Step { get; set; } = 1;
         public bool Pad { get; set; } = false;
